complete package structure, factor out mediapipe and video capture initialization
