+++ infrastructure/mediapipe/hands_initializer.py
@@ -1,0 +1,10 @@
+import mediapipe as mp
+
+
+def initialize_mediapipe_hands(arguments):
+    return mp.solutions.hands.Hands(
+        static_image_mode=arguments.use_static_image_mode,
+        max_num_hands=2,
+        min_detection_confidence=arguments.min_detection_confidence,
+        min_tracking_confidence=arguments.min_tracking_confidence,
+    )
