+++ infrastructure/openCV/draw/draw_debug_messages.py
