+++ infrastructure/openCV/draw/draw_hand_landmarks.py
