+++ infrastructure/openCV/video_capture/video_capture_initializer.py
@@ -1,0 +1,5 @@
+def initialize_video_capture(cv, arguments):
+    capture = cv.VideoCapture(arguments.device)
+    capture.set(cv.CAP_PROP_FRAME_WIDTH, arguments.width)
+    capture.set(cv.CAP_PROP_FRAME_HEIGHT, arguments.height)
+    return capture
